+++ R/package_helpers.R
@@ -92,3 +92,27 @@ basic_package_sticker <- function(package_name, brackets = TRUE) {
   ### set sticker
   usethis::use_logo(img = filename)
 }
+
+#' Handling of an unknown error
+#'
+#' @description
+#' This function reacts to an unknown error by throwing an error and linking
+#' to a GitHub issues site with the request to submit an issue.
+#'
+#' @param issue_link
+#' A \code{character}, the URL to a GitHub issues site.
+#'
+#' @return
+#' No return value, but it throws an error.
+
+unknown_error <- function(issue_link = "https://github.com/loelschlaeger/oeli/issues") {
+  checkmate::assert_string(
+    issue_link, pattern = "^https://github.com/[[:alpha:]]*/[[:alpha:]]*/issues$"
+  )
+  msg <- c(
+    "We are sorry, an unknown error occured.",
+    "i" = paste("Please submit an issue here:", cli::style_hyperlink(issue_link, issue_link))
+  )
+  cli::cli_abort(msg)
+}
+